Migrate interrupt and exception enums to riscv::interrupt
